v0.7.3 + v0.7.4 - Compressed so hard, the patches stuck together.
What's new?

I'm trying to cultivate a healthy relationship with my Copilot.
First I had to implement async migration with Polly retry policy,
because it was blocking synchronously in DraftService..
just like it did in ArticleService/Database earlier. I should know by now.

Also:
 - Got a healthcheck on Zipkin Mysql because depends-on is unavailable in swarm
 - Deleting a subscriber now directly uses ID in controller, as test sometimes had multiple
    Subscribers with the same email.
 - A bug with the SubscriberRepository, where it would return the deleted entity as it was,
    because it referred directly to the modified entity which still existed in the context past
    the ExecuteDeleteAsync method.

For the 0.7.4 compression has been implemented to Redis payload for Articles.
This is of course only for articles travelling across the network, as we
don't expect compression to do anything good for inmemory cached articles.
This is the second part of our Green Software Architecture, the first being
v0.7.1 with the implementation of "Fetch Data from Proximity", represented
by the inmemory cache.