Very small change to v0.7.4 - Not enough for me to ask someone to write patch notes
What's.. new?

Just forgot to implement compression for UpdateArticleAsync, and documented why
I'm not going to do anything for GetArticles. Seeing as the latter gets ALL
articles in the repository, the hope is that it's mostly used for test cases
and other purposes in staging, and never ever ever in production.
Still, it's in there somewhere.