v0.7.5 - "In the grim darkness of distributed systems, there is only bash"
What's new?

There used to be powershell scripts for managing tests.
Because powershell parsing is an absolute nightmare,
every script has been converted to bash.
This eases rapid deployment and testing, whether docker compose or docker swarm is used.
As always, docker compose alone does not star all required features
(as feature flags cannot be toggled during runtime),
but if the need arises, the option remains.
Scripts should already be executeable, but in a bind,
remember to use chmod -x.
So long.