v0.8.0 - Short but sweetly ServiceToggleMiddleware
What's new?

This is very short for breaking changes, but with the addition of the
AdminController in SubscriberService, the feature flag finally works
as intended. It now exposes endpoints bypassing middleware to directly
influence the feature flag on runtime. This is volatile, and as so
the service will always read from the config on restart/startup.
Be ware.

So Long.